new color management design starts to take shape (canvas color only for now)
git-svn-id: svn://localhost/trunk/ardour2@326 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ gtk2_ardour/crossfade_edit.cc
@@ -47,6 +47,7 @@
 #include "canvas_impl.h"
 #include "simplerect.h"
 #include "waveview.h"
+#include "color.h"
 
 using namespace std;
 using namespace ARDOUR;
@@ -128,7 +129,7 @@ CrossfadeEditor::CrossfadeEditor (Session& s, Crossfade& xf, double my, double m
 	toplevel->property_x2() =  10.0;
 	toplevel->property_y2() =  10.0;
 	toplevel->property_fill() =  true;
-	toplevel->property_fill_color_rgba() =  (guint32) color_map[cCrossfadeEditorBase];
+	toplevel->property_fill_color_rgba() = color_map[cCrossfadeEditorBase];
 	toplevel->property_outline_pixels() =  0;
 	toplevel->signal_event().connect (mem_fun (*this, &CrossfadeEditor::canvas_event));
